+++ PRN232Project/BusinessObjects/ProductImage.cs
@@ -8,14 +8,14 @@ namespace BusinessObjects
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [Required]
         [MaxLength(256)]
         public string Image { get; set; }
 
         [Required]
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
 
         [ForeignKey("ProductId")]
         public Product Product { get; set; }
